add specifi compile for macos
